New : Improved the code quality a little.
Also included training tutorials for yolox and yolo-nas so that anyone can train models on their own dataset.

This commit also includes the v2 of the weapon detectors for both yolo-nas and yolox. These models have been trained on the v2 of the weapon_detection dataset we have internaly curated